Chore: check for both delete modals if amount of reports is not 0

+++ src/app/test/test.component.ts
@@ -207,7 +207,7 @@ export class TestComponent implements OnInit {
 
   openDeleteModal(deleteAllReports: boolean): void {
     const reportsToBeDeleted: TestListItem[] = this.getSelectedReports();
-    if (reportsToBeDeleted.length > 0 || deleteAllReports) {
+    if (this.reports && (reportsToBeDeleted.length > 0 || (deleteAllReports && this.reports.length > 0))) {
       this.deleteModal.open(reportsToBeDeleted, deleteAllReports);
     }
   }
